+++ lib/src/constants.dart
@@ -1,4 +1,4 @@
-const version = '2.3.0';
+const version = '2.4.0';
 const configJsonName = 'config_v5.json';
 final DateTime distantPast = DateTime.utc(1970, 01, 01);
 final DateTime distantFuture =