+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.4.0 - 
+### Added
+- New `getAllValueDetails()` method.
+
+### Changed
+- Deprecated `getVariationId()` and `getAllVariationIds()` methods in favor of `getValueDetails()` and `getAllValueDetails()`.
+
 ## 2.3.0 - 2022-10-18
 ### Changed
 - Renamed `mode` configuration option to `pollingMode`.